mkdir src/lib/storage
Split storage.ts into storage/index.ts and storage/tweet.ts
